refactor(bench): move the harbor adapter's metrics cluster to its own module
`stella_harbor/__init__.py` sat at exactly 2150 lines — its own ratchet
ceiling, zero headroom — so the next PR to add a net line there failed the
gate before it could be reviewed. That is not hypothetical: #2136 arrived
with two genuine features and went red on `file size ratchet` alone.

The metrics cluster is the natural seam. `_sum_step_usage`,
`_valid_nonnegative_number`, `_valid_nonnegative_integer`,
`_extract_metrics`, `_load_json_object` and `_envelope_score` are one
closed unit: they parse Stella's result envelope and touch nothing else in
the adapter. Moved verbatim — no behavior change — with the four names the
adapter and `tests/test_adapter.py` still reach for re-exported from the
package, following `.posture`'s underscore convention. `math` was only ever
imported for `_valid_nonnegative_number` and leaves with it.

`__init__.py` drops 2150 → 1997, restoring ~150 lines of headroom.

The new module is registered in `_FIXED_ADAPTER_SOURCE_PATHS`: that tuple
is enumerated rather than globbed precisely so an adapter file cannot
execute without being byte-compared against the published commit, and it
fails closed. Its one added line is what pushes `secure_launcher.py` from
4297 to 4298 — the documented irreducible case (a required registration
entry in an already-oversized file), not a ceiling raised to go green.

The baseline diff is deliberately scoped to these two files. A full
`--update` also retightened `crates/stella-core/src/driver/tests.rs` by 297
lines from a merge this branch had no part in; adopting unrelated
shrinkage while stella-core PRs are in flight is the repo's most frequent
red-main cause (#2004), so that hunk was dropped.

Pure refactor: no witness test, and no test was deleted.

Closes #2161
